### 2024-01-27 #6289 @editorialbot
[PRE REVIEW]: ReLax: Efficient and Scalable Recourse Explanation Benchmarking using JAX
**Submitting author:** @BirkhoffG (<a href="http://orcid.org/0009-0000-6277-9003">Hangzhi Guo</a>)
**Repository:** https://github.com/BirkhoffG/jax-relax/
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.2.4
