### 2024-02-20 0ffc980
Implement `GumbelSoftmaxTransformation`; deprecate `OneHotTransformation`
