+++ relax/_modidx.py
@@ -182,6 +182,14 @@ d = { 'settings': { 'branch': 'master',
                                                                                       'relax/data_utils.py'),
                                   'relax.data_utils.FeaturesList.with_transformed_data': ( 'data.utils.html#featureslist.with_transformed_data',
                                                                                            'relax/data_utils.py'),
+                                  'relax.data_utils.GumbelSoftmaxTransformation': ( 'data.utils.html#gumbelsoftmaxtransformation',
+                                                                                    'relax/data_utils.py'),
+                                  'relax.data_utils.GumbelSoftmaxTransformation.__init__': ( 'data.utils.html#gumbelsoftmaxtransformation.__init__',
+                                                                                             'relax/data_utils.py'),
+                                  'relax.data_utils.GumbelSoftmaxTransformation.apply_constraints': ( 'data.utils.html#gumbelsoftmaxtransformation.apply_constraints',
+                                                                                                      'relax/data_utils.py'),
+                                  'relax.data_utils.GumbelSoftmaxTransformation.soft_constraints': ( 'data.utils.html#gumbelsoftmaxtransformation.soft_constraints',
+                                                                                                     'relax/data_utils.py'),
                                   'relax.data_utils.IdentityTransformation': ( 'data.utils.html#identitytransformation',
                                                                                'relax/data_utils.py'),
                                   'relax.data_utils.IdentityTransformation.__init__': ( 'data.utils.html#identitytransformation.__init__',
@@ -221,14 +229,6 @@ d = { 'settings': { 'branch': 'master',
                                   'relax.data_utils.OneHotEncoder.transform': ( 'data.utils.html#onehotencoder.transform',
                                                                                 'relax/data_utils.py'),
                                   'relax.data_utils.OneHotTransformation': ('data.utils.html#onehottransformation', 'relax/data_utils.py'),
-                                  'relax.data_utils.OneHotTransformation.__init__': ( 'data.utils.html#onehottransformation.__init__',
-                                                                                      'relax/data_utils.py'),
-                                  'relax.data_utils.OneHotTransformation.apply_constraints': ( 'data.utils.html#onehottransformation.apply_constraints',
-                                                                                               'relax/data_utils.py'),
-                                  'relax.data_utils.OneHotTransformation.compute_reg_loss': ( 'data.utils.html#onehottransformation.compute_reg_loss',
-                                                                                              'relax/data_utils.py'),
-                                  'relax.data_utils.OneHotTransformation.num_categories': ( 'data.utils.html#onehottransformation.num_categories',
-                                                                                            'relax/data_utils.py'),
                                   'relax.data_utils.OrdinalPreprocessor': ('data.utils.html#ordinalpreprocessor', 'relax/data_utils.py'),
                                   'relax.data_utils.OrdinalPreprocessor.fit': ( 'data.utils.html#ordinalpreprocessor.fit',
                                                                                 'relax/data_utils.py'),
@@ -242,6 +242,10 @@ d = { 'settings': { 'branch': 'master',
                                                                                        'relax/data_utils.py'),
                                   'relax.data_utils.OrdinalTransformation.num_categories': ( 'data.utils.html#ordinaltransformation.num_categories',
                                                                                              'relax/data_utils.py'),
+                                  'relax.data_utils.SoftmaxTransformation': ( 'data.utils.html#softmaxtransformation',
+                                                                              'relax/data_utils.py'),
+                                  'relax.data_utils.SoftmaxTransformation.soft_constraints': ( 'data.utils.html#softmaxtransformation.soft_constraints',
+                                                                                               'relax/data_utils.py'),
[…]
