+++ relax/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.2.5"
+__version__ = "0.2.6"
 
 from .data_module import DataModule, DataModuleConfig, load_data
 from .data_utils import(
