+++ settings.ini
@@ -1,7 +1,7 @@
 [DEFAULT]
 repo = jax-relax
 lib_name = jax-relax
-version = 0.2.5
+version = 0.2.6
 min_python = 3.9
 license = apache2
 black_formatting = False
