### 2024-02-22 f89c059
Fix `GumbelSoftmaxTransformation.name`; add more tests for set_transformation
