+++ relax/_modidx.py
@@ -244,6 +244,8 @@ d = { 'settings': { 'branch': 'master',
                                                                                              'relax/data_utils.py'),
                                   'relax.data_utils.SoftmaxTransformation': ( 'data.utils.html#softmaxtransformation',
                                                                               'relax/data_utils.py'),
+                                  'relax.data_utils.SoftmaxTransformation.__init__': ( 'data.utils.html#softmaxtransformation.__init__',
+                                                                                       'relax/data_utils.py'),
                                   'relax.data_utils.SoftmaxTransformation.soft_constraints': ( 'data.utils.html#softmaxtransformation.soft_constraints',
                                                                                                'relax/data_utils.py'),
                                   'relax.data_utils.Transformation': ('data.utils.html#transformation', 'relax/data_utils.py'),
