### 2024-03-31 #6289 @GarrettMerz
Hi, all-

I'm happy to help! I'm slightly less familiar with JAX compared to Pytorch/TF, so it might be optimal to have at least one other set of eyes on this in case I miss anything, but I'm happy to review!

-Garrett
