### 2024-03-31 #6289 @duhd1993
I'd be happy to take a look. I have experience with JAX
