### 2024-04-03 f8b5618
Update GumbelSoftmaxTransformation tau value and add to_dict method
