+++ relax/_modidx.py
@@ -190,6 +190,8 @@ d = { 'settings': { 'branch': 'master',
                                                                                                       'relax/data_utils.py'),
                                   'relax.data_utils.GumbelSoftmaxTransformation.soft_constraints': ( 'data.utils.html#gumbelsoftmaxtransformation.soft_constraints',
                                                                                                      'relax/data_utils.py'),
+                                  'relax.data_utils.GumbelSoftmaxTransformation.to_dict': ( 'data.utils.html#gumbelsoftmaxtransformation.to_dict',
+                                                                                            'relax/data_utils.py'),
                                   'relax.data_utils.IdentityTransformation': ( 'data.utils.html#identitytransformation',
                                                                                'relax/data_utils.py'),
                                   'relax.data_utils.IdentityTransformation.__init__': ( 'data.utils.html#identitytransformation.__init__',
