### 2024-04-03 83f01dc
Update requirements in settings.ini
