+++ settings.ini
@@ -27,7 +27,7 @@ keywords = JAX, Recourse, Explanation, Interpretability, Machine Learning
 language = English
 status = 3
 user = birkhoffg
-requirements = scikit-learn>=1.0.2,<1.4.0 pandas jax[cpu] tqdm optax pydantic>=1.9.0,<2 jax-tqdm einops keras>=3.0.3 matplotlib seaborn dm-haiku
+requirements = scikit-learn>=1.0.2 pandas jax[cpu] tqdm optax pydantic>=1.9.0,<2 jax-tqdm einops keras>=3.0.3 matplotlib seaborn dm-haiku
 dev_requirements = torch>=1.7.0 causalgraphicalmodels pre-commit datasets nbdev jupyter
 readme_nb = index.ipynb
 allowed_metadata_keys = 
