### 2024-04-03 6fdc928
[WIP] Refactor data_utils
