+++ relax/_modidx.py
@@ -286,6 +286,90 @@ d = { 'settings': { 'branch': 'master',
                                                                                                'relax/data_utils.py'),
                                   'relax.data_utils._check_xs': ('data.utils.html#_check_xs', 'relax/data_utils.py'),
                                   'relax.data_utils._unique': ('data.utils.html#_unique', 'relax/data_utils.py')},
+            'relax.data_utils.preprocessing': { 'relax.data_utils.preprocessing.DataPreprocessor': ( 'data_utils/preprocessing.html#datapreprocessor',
+                                                                                                     'relax/data_utils/preprocessing.py'),
+                                                'relax.data_utils.preprocessing.DataPreprocessor.__init__': ( 'data_utils/preprocessing.html#datapreprocessor.__init__',
+                                                                                                              'relax/data_utils/preprocessing.py'),
+                                                'relax.data_utils.preprocessing.DataPreprocessor.fit': ( 'data_utils/preprocessing.html#datapreprocessor.fit',
+                                                                                                         'relax/data_utils/preprocessing.py'),
+                                                'relax.data_utils.preprocessing.DataPreprocessor.fit_transform': ( 'data_utils/preprocessing.html#datapreprocessor.fit_transform',
+                                                                                                                   'relax/data_utils/preprocessing.py'),
+                                                'relax.data_utils.preprocessing.DataPreprocessor.from_dict': ( 'data_utils/preprocessing.html#datapreprocessor.from_dict',
+                                                                                                               'relax/data_utils/preprocessing.py'),
+                                                'relax.data_utils.preprocessing.DataPreprocessor.inverse_transform': ( 'data_utils/preprocessing.html#datapreprocessor.inverse_transform',
+                                                                                                                       'relax/data_utils/preprocessing.py'),
+                                                'relax.data_utils.preprocessing.DataPreprocessor.to_dict': ( 'data_utils/preprocessing.html#datapreprocessor.to_dict',
+                                                                                                             'relax/data_utils/preprocessing.py'),
+                                                'relax.data_utils.preprocessing.DataPreprocessor.transform': ( 'data_utils/preprocessing.html#datapreprocessor.transform',
+                                                                                                               'relax/data_utils/preprocessing.py'),
+                                                'relax.data_utils.preprocessing.EncoderPreprocessor': ( 'data_utils/preprocessing.html#encoderpreprocessor',
+                                                                                                        'relax/data_utils/preprocessing.py'),
+                                                'relax.data_utils.preprocessing.EncoderPreprocessor._fit': ( 'data_utils/preprocessing.html#encoderpreprocessor._fit',
+                                                                                                             'relax/data_utils/preprocessing.py'),
+                                                'relax.data_utils.preprocessing.EncoderPreprocessor._inverse_transform': ( 'data_utils/preprocessing.html#encoderpreprocessor._inverse_transform',
+                                                                                                                           'relax/data_utils/preprocessing.py'),
+                                                'relax.data_utils.preprocessing.EncoderPreprocessor._transform': ( 'data_utils/preprocessing.html#encoderpreprocessor._transform',
+                                                                                                                   'relax/data_utils/preprocessing.py'),
+                                                'relax.data_utils.preprocessing.EncoderPreprocessor.from_dict': ( 'data_utils/preprocessing.html#encoderpreprocessor.from_dict',
+                                                                                                                  'relax/data_utils/preprocessing.py'),
+                                                'relax.data_utils.preprocessing.EncoderPreprocessor.to_dict': ( 'data_utils/preprocessing.html#encoderpreprocessor.to_dict',
+                                                                                                                'relax/data_utils/preprocessing.py'),
+                                                'relax.data_utils.preprocessing.MinMaxScaler': ( 'data_utils/preprocessing.html#minmaxscaler',
[…]
