### 2024-04-03 fc7e560
Finished refactor; Fix bug in with_transformation
