+++ relax/data_utils/features.py
@@ -1,324 +1,17 @@
-# AUTOGENERATED! DO NOT EDIT! File to edit: ../nbs/01_data.utils.ipynb.
+# AUTOGENERATED! DO NOT EDIT! File to edit: ../../nbs/data_utils/features.ipynb.
 
-# %% ../nbs/01_data.utils.ipynb 3
+# %% ../../nbs/data_utils/features.ipynb 3
 from __future__ import annotations
-from fastcore.test import *
-import pandas as pd
-import numpy as np
-import jax
-import jax.numpy as jnp
-import einops
-import os, sys, json, pickle
-import shutil
-from .utils import gumbel_softmax, load_pytree, save_pytree, get_config
-import chex
-import functools as ft
-import warnings
+from .preprocessing import *
+from .transforms import *
+from ..utils import get_config, gumbel_softmax, load_pytree, save_pytree
+from ..import_essentials import *
+from copy import deepcopy
 
 # %% auto 0
-__all__ = ['PREPROCESSING_TRANSFORMATIONS', 'DataPreprocessor', 'MinMaxScaler', 'EncoderPreprocessor', 'OrdinalPreprocessor',
-           'OneHotEncoder', 'Transformation', 'MinMaxTransformation', 'SoftmaxTransformation',
-           'GumbelSoftmaxTransformation', 'OneHotTransformation', 'OrdinalTransformation', 'IdentityTransformation',
-           'Feature', 'FeaturesList']
+__all__ = ['Feature', 'FeaturesList']
 
-# %% ../nbs/01_data.utils.ipynb 6
-def _check_xs(xs: np.ndarray, name: str):
-    if xs.ndim > 2 or (xs.ndim == 2 and xs.shape[1] != 1):
-        raise ValueError(f"`{name}` only supports array with a single feature, but got shape={xs.shape}.")
-    
-        
-class DataPreprocessor:
-
-    def __init__(
-        self, 
-        name: str = None # The name of the preprocessor. If None, the class name will be used.
-    ):
-        """Base class for data preprocessors."""
-        self.name = name or self.__class__.__name__
-    
-    def fit(self, xs, y=None):
-        """Fit the preprocessor with `xs` and `y`."""
-        raise NotImplementedError
-    
-    def transform(self, xs):
-        """Transform `xs`."""
-        raise NotImplementedError
-    
-    def fit_transform(self, xs, y=None):
-        """Fit the preprocessor with `xs` and `y`, then transform `xs`."""
-        self.fit(xs, y)
-        return self.transform(xs)
-    
-    def inverse_transform(self, xs):
-        """Inverse transform `xs`."""
-        raise NotImplementedError
-    
-    def to_dict(self) -> dict:
-        """Convert the preprocessor to a dictionary."""
-        raise NotImplementedError
-    
-    def from_dict(self, params: dict):
-        """Load the attributes of the preprocessor from a dictionary."""
-        raise NotImplementedError
-        
-    __ALL__ = ["fit", "transform", "fit_transform", "inverse_transform", "to_dict", "from_dict"]
-
-# %% ../nbs/01_data.utils.ipynb 7
-class MinMaxScaler(DataPreprocessor): 
-    def __init__(self):
-        super().__init__(name="minmax")
-        
-    def fit(self, xs, y=None):
-        _check_xs(xs, name="MinMaxScaler")
-        self.min_ = xs.min(axis=0)
-        self.max_ = xs.max(axis=0)
-        return self
-    
-    def transform(self, xs):
-        return (xs - self.min_) / (self.max_ - self.min_)
-    
-    def inverse_transform(self, xs):
-        return xs * (self.max_ - self.min_) + self.min_
-    
[…]
