+++ relax/_modidx.py
@@ -83,209 +83,94 @@ d = { 'settings': { 'branch': 'master',
                                    'relax.data_module.features2pandas': ('data.html#features2pandas', 'relax/data_module.py'),
                                    'relax.data_module.load_data': ('data.html#load_data', 'relax/data_module.py'),
                                    'relax.data_module.to_feature': ('data.html#to_feature', 'relax/data_module.py')},
-            'relax.data_utils': { 'relax.data_utils.DataPreprocessor': ('data.utils.html#datapreprocessor', 'relax/data_utils.py'),
-                                  'relax.data_utils.DataPreprocessor.__init__': ( 'data.utils.html#datapreprocessor.__init__',
-                                                                                  'relax/data_utils.py'),
-                                  'relax.data_utils.DataPreprocessor.fit': ('data.utils.html#datapreprocessor.fit', 'relax/data_utils.py'),
-                                  'relax.data_utils.DataPreprocessor.fit_transform': ( 'data.utils.html#datapreprocessor.fit_transform',
-                                                                                       'relax/data_utils.py'),
-                                  'relax.data_utils.DataPreprocessor.from_dict': ( 'data.utils.html#datapreprocessor.from_dict',
-                                                                                   'relax/data_utils.py'),
-                                  'relax.data_utils.DataPreprocessor.inverse_transform': ( 'data.utils.html#datapreprocessor.inverse_transform',
-                                                                                           'relax/data_utils.py'),
-                                  'relax.data_utils.DataPreprocessor.to_dict': ( 'data.utils.html#datapreprocessor.to_dict',
-                                                                                 'relax/data_utils.py'),
-                                  'relax.data_utils.DataPreprocessor.transform': ( 'data.utils.html#datapreprocessor.transform',
-                                                                                   'relax/data_utils.py'),
-                                  'relax.data_utils.EncoderPreprocessor': ('data.utils.html#encoderpreprocessor', 'relax/data_utils.py'),
-                                  'relax.data_utils.EncoderPreprocessor._fit': ( 'data.utils.html#encoderpreprocessor._fit',
-                                                                                 'relax/data_utils.py'),
-                                  'relax.data_utils.EncoderPreprocessor._inverse_transform': ( 'data.utils.html#encoderpreprocessor._inverse_transform',
-                                                                                               'relax/data_utils.py'),
-                                  'relax.data_utils.EncoderPreprocessor._transform': ( 'data.utils.html#encoderpreprocessor._transform',
-                                                                                       'relax/data_utils.py'),
-                                  'relax.data_utils.EncoderPreprocessor.from_dict': ( 'data.utils.html#encoderpreprocessor.from_dict',
-                                                                                      'relax/data_utils.py'),
-                                  'relax.data_utils.EncoderPreprocessor.to_dict': ( 'data.utils.html#encoderpreprocessor.to_dict',
-                                                                                    'relax/data_utils.py'),
-                                  'relax.data_utils.Feature': ('data.utils.html#feature', 'relax/data_utils.py'),
-                                  'relax.data_utils.Feature.__get_item__': ('data.utils.html#feature.__get_item__', 'relax/data_utils.py'),
-                                  'relax.data_utils.Feature.__init__': ('data.utils.html#feature.__init__', 'relax/data_utils.py'),
-                                  'relax.data_utils.Feature.__repr__': ('data.utils.html#feature.__repr__', 'relax/data_utils.py'),
-                                  'relax.data_utils.Feature._dispatch_transformation': ( 'data.utils.html#feature._dispatch_transformation',
-                                                                                         'relax/data_utils.py'),
-                                  'relax.data_utils.Feature._init_is_categorical': ( 'data.utils.html#feature._init_is_categorical',
[…]
