### 2024-04-04 1cb16be
Update type hint
