### 2024-04-27 9928e3b
Fix bug when GrowingSphere generates cfs without constraints
