### 2024-05-28 c722a81
update installation instructions
