### 2024-05-28 7a56e65
Use jnp.clip for applying constraints
