### 2024-05-28 5e39cbc
Delete legacy TabularDataModule
