+++ relax/_modidx.py
@@ -66,13 +66,6 @@ d = { 'settings': { 'branch': 'master',
                                                                                             'relax/data_module.py'),
                                    'relax.data_module.DataModuleInfoMixin.xs': ('data.html#datamoduleinfomixin.xs', 'relax/data_module.py'),
                                    'relax.data_module.DataModuleInfoMixin.ys': ('data.html#datamoduleinfomixin.ys', 'relax/data_module.py'),
-                                   'relax.data_module.TabularDataModule': ('data.html#tabulardatamodule', 'relax/data_module.py'),
-                                   'relax.data_module.TabularDataModule.__init__': ( 'data.html#tabulardatamodule.__init__',
-                                                                                     'relax/data_module.py'),
-                                   'relax.data_module.TabularDataModuleConfigs': ( 'data.html#tabulardatamoduleconfigs',
-                                                                                   'relax/data_module.py'),
-                                   'relax.data_module.TabularDataModuleConfigs.__ini__': ( 'data.html#tabulardatamoduleconfigs.__ini__',
-                                                                                           'relax/data_module.py'),
                                    'relax.data_module._validate_dataname': ('data.html#_validate_dataname', 'relax/data_module.py'),
                                    'relax.data_module.dataframe2features': ('data.html#dataframe2features', 'relax/data_module.py'),
                                    'relax.data_module.dataframe2labels': ('data.html#dataframe2labels', 'relax/data_module.py'),
