### 2024-05-29 #6567 @BirkhoffG
@GarrettMerz Thank you very much for your constructive feedback and suggestions. Below, we clarify your concerns about our library:

> -The "installation instructions" are incomplete. I'd like some more clarity on how and when to install JAX (the repo says to install jax-relax before the GPU version of JAX- is this correct?). Is this something that can be clarified? Additionally, several required packages are not listed: fastcore, nbdev, ipywidgets>=7.0.

Our latest installation guidance clarifies how to install JAX with GPU (see the latest [Installation](https://birkhoffg.github.io/jax-relax/tutorials/install.html) section). Essentially, it is advised to first install `jax-relax` (i.e., `pip install jax-relax`), then install the GPU version of JAX according to the [JAX official installation instructions](https://jax.readthedocs.io/en/latest/installation.html).

Additionally, these mentioned packages (including fastcore, nbdev) are not required for using `jax-relax` as an end-user. For example, you do not need these packages if you run code examples in the [getting started](https://birkhoffg.github.io/jax-relax/tutorials/getting_started.html) section. On the other hand, if you are a contributor to `jax-relax`, you do need these packages, and you should install them via `pip install "jax-relax[dev]"`, which will automatically install `jax-relax` and all required packages for development. We have updated the docs to reflect those requirements.

> -Much of the documentation is located on the linked doc page rather than in the repository. This is fine, especially for things like comprehensive API documentation, but the "community guidelines" section (or a link to it) should be in the main repository.

All of the documentations are written in Jupyter notebook, which is automatically converted to the documentation web pages. Those Jupyter notebooks are located in the GitHub repository. For example, the "community guidelines" section is located at https://github.com/BirkhoffG/jax-relax/blob/master/nbs/tutorials/contribution.ipynb.

> -The repository does not have automated tests. However, there is a comprehensive tutorial that describes how to use the software. The tutorial is somewhat spread across the repo- the steps in the README give a way to test a sample classifier, but the 'Getting Started' section refers to the tutorial notebooks. If possible, this should be centralized and streamlined more- if I can test that everything runs out-of-the-box by clicking through one or more notebooks, that's ideal.

Our repository contains automated tests by simply running `nbdev_test`, which is enabled in the continuous integration (CI) via the GitHub Action. See https://github.com/BirkhoffG/jax-relax/blob/c722a81bbbd3012bcd46a9154529303302c604e5/.github/workflows/test.yaml#L85-L86

All test cases are located in each Jupyter notebook. Essentially, all code blocks without directives are considered test cases (directives are special comments start with `#|`, e.g., `#| export` is the most common directive. See [this](https://nbdev.fast.ai/explanations/directives.html)). Refer to [this section](https://birkhoffg.github.io/jax-relax/tutorials/contribution.html#write-test-cases-in-jupyter-notebook) on how we write unit tests.

In terms of the tutorials, the [getting started](https://birkhoffg.github.io/jax-relax/tutorials/getting_started.html) and [`ReLax` as a Recourse Library](https://birkhoffg.github.io/jax-relax/tutorials/methods.html) are supposed to be the superset of the contents covered in [README](https://github.com/BirkhoffG/jax-relax/blob/master/README.md), i.e., these two tutorials should cover everything explained in the [Dive into ReLax](https://github.com/BirkhoffG/jax-relax?tab=readme-ov-file#dive-into-relax) section. If we miss something, please let us know, and we will update the tutorials.

> -The datasets needed to reproduce the plots in the paper are not provided in the repository, but are linked in the paper. When links are provided, as with the 'adult' dataset mentioned in the tutorials notebook, the available data is not formatted as needed by the code. I'd like to reproduce the benchmarks in the paper, but it's unclear what I need to do in order to get a hold of the data to do so.

Due to the concerns about the [file size limit in GitHub](https://docs.github.com/en/repositories/working-with-files/managing-large-files/about-large-files-on-github#file-size-limits), we choose to host datasets and models on [huggingface](https://huggingface.co/datasets/birkhoffg/ReLax-Assets). All datasets are located in `https://huggingface.co/datasets/birkhoffg/ReLax-Assets/tree/main/{data_name}/data/data.csv` (replace `{data_name}` to the real dataset name, such as "adult").

We provide an easy-to-use API to automatically load the data modules VIA `load_data`:
```python
import relax

data = relax.load_data('adult')
```

To reproduce the results in the paper, you can run
```bash
python -m benchmarks.built-in.run_all
```

cc @Fei-Tao
