### 2024-08-09 #6567 @duhd1993
Sorry for the delay. I have review the paper and the Github repo.  @BirkhoffG

ReLax is claimed to be a JAX-based library for efficient and scalable benchmarking of recourse and counterfactual explanation methods for machine learning models. It implements 9 popular recourse methods and includes 14 medium-sized datasets and 1 large-scale dataset for benchmarking. ReLax achieves significant performance improvements.

There are a few issues that needs improvement or clarifications

- The installation instruction is not complete. When I try to reproduce the results. There are errors for missing libraries `nbdev,fastcore`. I think you need to write a requirements.txt and test from a clean environement.
- Figure 3 demonstrated the efficiency on large scale dataset. But how do you ensure the quality and correctness of results given the lack of baselines at that scale?
- Is there a review on most popular recourse methods? Why you choose the existing ones? How easy is it to implement new recourse methods in ReLax? Is there a standardized interface?
