### 2024-08-26 8031e60
Update Python versions in setup.py
