+++ .github/workflows/test.yaml
@@ -15,7 +15,7 @@ jobs:
       - name: Set up Python
         uses: actions/setup-python@v4
         with:
-          python-version: "3.9"
+          python-version: "3.10"
           cache: "pip"
           cache-dependency-path: settings.ini
 
@@ -54,11 +54,11 @@ jobs:
     strategy:
       fail-fast: true
       matrix:
-        py: ['3.9', '3.10', '3.11']
+        py: ['3.9', '3.10', '3.11', '3.12']
     runs-on: ubuntu-latest
     steps:
       - name: Checkout Code
-        uses: actions/checkout@v3
+        uses: actions/checkout@v4
         # with:
         #   submodules: recursive
 
@@ -66,7 +66,7 @@ jobs:
         run: git submodule update --init --recursive
 
       - name: Set up Python
-        uses: actions/setup-python@v4
+        uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.py }}
           cache: "pip"
