### 2024-08-28 fa50357
Add instructions for benchmarking
