+++ benchmarks/README.md
@@ -1,0 +1,18 @@
+# Benchmarking Scripts
+
+This directory hosts scripts for benchmarking recourse methods.
+
+## Installation
+
+To install the necessary dependencies, run the following command:
+```sh
+pip install "jax-relax[dev]"
+```
+
+## Usage
+
+To reproduce the results in the paper, you can run
+
+```sh
+python -m benchmarks.built-in.run_all
+```
