+++ relax/utils.py
@@ -1,3 +1,5 @@
+"""Define utility funtions for `relax`."""
+
 # AUTOGENERATED! DO NOT EDIT! File to edit: ../nbs/00_utils.ipynb.
 
 # %% ../nbs/00_utils.ipynb 3
