+++ relax/data_module.py
@@ -1,3 +1,5 @@
+"""`DataModule` for training parametric models, generating and benchmarking CF explanations."""
+
 # AUTOGENERATED! DO NOT EDIT! File to edit: ../nbs/01_data.ipynb.
 
 # %% ../nbs/01_data.ipynb 3
