+++ relax/methods/counternet.py
@@ -1,3 +1,5 @@
+"""A prediction-aware recourse model"""
+
 # AUTOGENERATED! DO NOT EDIT! File to edit: ../../nbs/methods/04_counternet.ipynb.
 
 # %% ../../nbs/methods/04_counternet.ipynb 3
