+++ relax/legacy/ckpt_manager.py
@@ -1,3 +1,5 @@
+"""Manage the model and optimizer checkpoints"""
+
 # AUTOGENERATED! DO NOT EDIT! File to edit: ../../nbs/legacy/04b_ckpt_manager.ipynb.
 
 # %% auto 0
