### 2024-08-28 f2e5b4c
Add cost-invalidity trade-off analysis to paper
