### 2024-08-28 44d5a36
Update legacy benchmark scripts and dependencies
