### 2024-08-29 #6567 @BirkhoffG
@duhd1993 Thank you very much for your constructive feedback and suggestions. Below, we clarify your concerns about our library:

> The installation instruction is not complete. When I tried to reproduce the results, there are errors for missing libraries `nbdev, fastcore`. I think you need to write a requirements.txt and test from a clean environment.
>

As we suggested [here](https://github.com/openjournals/joss-reviews/issues/6567#issuecomment-2136293116), we deliberately exclude these requirements (including fastcore, nbdev) for normal installation because these packages are not required for using `jax-relax` as an end-user. For example, you do not need these packages if you run code examples in the [getting started](https://birkhoffg.github.io/jax-relax/tutorials/getting_started.html) section. These packages are only required if you are a contributor/developer to `jax-relax`. In that case, you should install them via `pip install "jax-relax[dev]"`.

We have updated the [installation docs](https://birkhoffg.github.io/jax-relax/tutorials/install.html) to reflect those requirements. The dependencies are specified in [settings.ini](https://github.com/BirkhoffG/jax-relax/blob/44d5a36debf8c1c8516ca71dd01c268ebd604fdc/settings.ini#L30-L31).

> Figure 3 demonstrated the efficiency on large scale dataset. But how do you ensure the quality and correctness of results given the lack of baselines at that scale?
>

You are right in pointing out that existing libraries/implementations cannot scale, and this is the main limitation that we aim to address in `jax-relax`. We try our best to implement each recourse method faithfully to their original papers and the official code provided (if provided). We evaluate the quality of the recourse methods in medium-sized datasets, and we use this implementation (evaluated correctly at a medium-sized scale) to test how these methods work at a larger scale at which there is no point of comparison for `jax-relax` simply because no other implementation of recourse methods scales up to such large-sized datasets.

Our experiments show a similar pattern between the results achieved by the implementations of recourse methods in `jax-relax` on the medium-sized datasets and the large-scale dataset (see this [figure](https://github.com/BirkhoffG/jax-relax/blob/3d1bd6e727942ddf3ba3c0218bdb8935ce817381/paper/figs/tradeoffs.pdf). We will update this result in the final paper.

> Is there a review on most popular recourse methods? Why you choose the existing ones?
>

There exists a large amount of recourse methods proposed by the research communities. These surveys [1, 2, 3] provide a comprehensive review of the various problems and methods for algorithmic recourses. Based on the use of parametric models, we categorize existing methods into non-parametric, semi-parametric, and parametric methods, and we select 3 representative methods for each category. Additionally, given the rapid progress in this field, it is impractical to incorporate every existing recourse method, and we welcome the research communities to contribute new recourse methods to `jax-relax`.

> How easy is it to implement new recourse methods in ReLax? Is there a standardized interface?
>

Yes, you can easily implement new recourse methods and integrate them into the `jax-relax` pipeline. Essentially, to implement a new recourse method, you need to ensure that it inherits from [`CFModule`](https://birkhoffg.github.io/jax-relax/methods/base.html#cfmodule) class. Check out our tutorials on [implementing your own recourse methods](https://birkhoffg.github.io/jax-relax/tutorials/methods.html#implement-your-own-recourse-methods).

> I tried to run the benchmark script and got this error: `ModuleNotFoundError: No module named 'relax.module'`
>

We have updated a [README](https://github.com/BirkhoffG/jax-relax/blob/master/benchmarks/README.md) on requirements and instructions on how to run the benchmark scripts.

## Reference

[1] https://arxiv.org/abs/2010.10596
[2] https://ieeexplore.ieee.org/abstract/document/9321372
[3] https://arxiv.org/abs/2010.04050
