+++ paper/paper.bib
@@ -1,1055 +1,91 @@
-% bi-level
-@inproceedings{shaban2019truncated,
-  title={Truncated back-propagation for bilevel optimization},
-  author={Shaban, Amirreza and Cheng, Ching-An and Hatch, Nathan and Boots, Byron},
-  booktitle={The 22nd International Conference on Artificial Intelligence and Statistics},
-  pages={1723--1732},
-  year={2019},
-  organization={PMLR}
-}
-
-@article{gdpr,
-author = {Chris Jay Hoofnagle and Bart van der Sloot and Frederik Zuiderveen Borgesius},
-title = {The European Union general data protection regulation: what it is and what it means},
-journal = {Information \& Communications Technology Law},
-volume = {28},
-number = {1},
-pages = {65-98},
-year  = {2019},
-publisher = {Routledge}
-
-}
-
-
-@article{gu2022min,
-  title={Min-Max Bilevel Multi-objective Optimization with Applications in Machine Learning},
-  author={Gu, Alex and Lu, Songtao and Ram, Parikshit and Weng, Lily},
-  journal={arXiv preprint arXiv:2203.01924},
-  year={2022}
-}
-
-
-% HPO
-@inproceedings{maclaurin2015gradient,
-  title={Gradient-based hyperparameter optimization through reversible learning},
-  author={Maclaurin, Dougal and Duvenaud, David and Adams, Ryan},
-  booktitle={International conference on machine learning},
-  pages={2113--2122},
-  year={2015},
-  organization={PMLR}
-}
-
-%German Credit
-@misc{asuncion2007uci, 
-  title={UCI machine learning repository},
-  author={Asuncion, Arthur and Newman, David},
-  year={2007},
-  publisher={Irvine, CA, USA}
-}
-
-% student
-@article{cortez2008using,
-  title={Using data mining to predict secondary school student performance},
-  author={Cortez, Paulo and Silva, Alice Maria Gon{\c{c}}alves},
-  year={2008},
-  publisher={EUROSIS-ETI}
-}
-
-
-%=========================Adv Training
-
-@article{goodfellow2014explaining,
-  title={Explaining and harnessing adversarial examples},
-  author={Goodfellow, Ian J and Shlens, Jonathon and Szegedy, Christian},
-  journal={arXiv preprint arXiv:1412.6572},
-  year={2014}
-}
-
-
-% backdoor
-@inproceedings{saha2020hidden,
-  title={Hidden trigger backdoor attacks},
-  author={Saha, Aniruddha and Subramanya, Akshayvarun and Pirsiavash, Hamed},
-  booktitle={Proceedings of the AAAI conference on artificial intelligence},
-  volume={34},
-  number={07},
-  pages={11957--11965},
-  year={2020}
-}
-% patch
-@article{gu2019badnets,
-  title={Badnets: Evaluating backdooring attacks on deep neural networks},
-  author={Gu, Tianyu and Liu, Kang and Dolan-Gavitt, Brendan and Garg, Siddharth},
-  journal={IEEE Access},
-  volume={7},
-  pages={47230--47244},
-  year={2019},
-  publisher={IEEE}
-}
-
-% Posioning attack: Meta Poison
-@article{huang2020metapoison,
-  title={Metapoison: Practical general-purpose clean-label data poisoning},
-  author={Huang, W Ronny and Geiping, Jonas and Fowl, Liam and Taylor, Gavin and Goldstein, Tom},
-  journal={Advances in Neural Information Processing Systems},
-  volume={33},
-  pages={12080--12091},
-  year={2020}
-}
-% Gradient Matching
-@inproceedings{
-geiping2021witches,
-title={Witches' Brew: Industrial Scale Data Poisoning via Gradient Matching},
[…]
