### 2024-10-20 2ea0071
update affiliations
