### 2024-10-20 #6567 @BirkhoffG
Hi @Fei-Tao,

I have completed the affiliations and references. Please let me know if I missed something.

I have also archived the release on Zenodo, and the DOI is this: https://doi.org/10.5281/zenodo.13957805
