+++ paper/paper.bib
@@ -19,6 +19,7 @@
   author={Wachter, Sandra and Mittelstadt, Brent and Russell, Chris},
   journal={Cybersecurity},
   year={2017},
+  doi={10.2139/ssrn.3063289},
   url={https://api.semanticscholar.org/CorpusID:3995299}
 }
 @article{verma2020counterfactual,
@@ -75,16 +76,17 @@
     series = {CHI '18}
 }
 @article{miller2019explanation,
-    title = {Explanation in artificial intelligence: Insights from the social sciences},
-    journal = {Artificial Intelligence},
-    volume = {267},
-    pages = {1-38},
-    year = {2019},
-    issn = {0004-3702},
-    doi = {https://doi.org/10.1016/j.artint.2018.07.007},
-    url = {https://www.sciencedirect.com/science/article/pii/S0004370218305988},
-    author = {Tim Miller},
-    keywords = {Explanation, Explainability, Interpretability, Explainable AI, Transparency},
+  author       = {Tim Miller},
+  title        = {Explanation in artificial intelligence: Insights from the social sciences},
+  journal      = {Artif. Intell.},
+  volume       = {267},
+  pages        = {1--38},
+  year         = {2019},
+  url          = {https://doi.org/10.1016/j.artint.2018.07.007},
+  doi          = {10.1016/J.ARTINT.2018.07.007},
+  timestamp    = {Thu, 25 May 2023 12:52:41 +0200},
+  biburl       = {https://dblp.org/rec/journals/ai/Miller19.bib},
+  bibsource    = {dblp computer science bibliography, https://dblp.org}
 }
 @inproceedings{Bhatt20explainable,
 author = {Bhatt, Umang and Xiang, Alice and Sharma, Shubham and Weller, Adrian and Taly, Ankur and Jia, Yunhan and Ghosh, Joydeep and Puri, Ruchir and Moura, Jos\'{e} M. F. and Eckersley, Peter},
