### 2024-10-22 #6567 @Fei-Tao
Hi @BirkhoffG the article and reference look good to me now.
Would you please make a new release of the software with the latest changes from the review and post the version number here? Thanks for your time.
