### 2024-10-22 #6567 @BirkhoffG
Hi @Fei-Tao
Our current latest release has addressed all the comments from the reviewers. The version number is [v0.2.8](https://github.com/BirkhoffG/jax-relax/releases/tag/v0.2.8).
