### 2024-10-22 #6567 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
