### 2024-11-12 #6567 @crvernon
:wave: @BirkhoffG - I just need you to address the following before I move to accept this submission for publication:

- [ ] LINE 112:  Curly brackets are showing up once rendered around "CLUE".  In your bib file, you have: `{Getting a {\{}CLUE{\}}: A  Method for Explaining Uncertainty Estimates}` which can be reduced to `{Getting a {CLUE}: A  Method for Explaining Uncertainty Estimates}` to fix this.
- [ ] LINE 156:  The "p" in "python" should be capitalized.  You can just wrap the "P" in your bib file with curly brackets like the previous bullet points out.

Let me know when these have been taken care of.  Thanks.
