+++ paper/paper.bib
@@ -186,7 +186,7 @@ location = {Birmingham, United Kingdom},
 series = {CIKM '23}
 }
 @article{pawelczyk2021carla,
-  title={CARLA: A Python Library to Benchmark Algorithmic Recourse and Counterfactual Explanation Algorithms},
+  title={CARLA: A {Python} Library to Benchmark Algorithmic Recourse and Counterfactual Explanation Algorithms},
   author={Martin Pawelczyk and Sascha Bielawski and Jan van den Heuvel and Tobias Richter and Gjergji Kasneci},
   journal={ArXiv},
   year={2021},
@@ -262,7 +262,7 @@ series = {CIKM '23}
     series = {WWW '20}
 }
 @inproceedings{antoran2021clue,
-title={Getting a {\{}CLUE{\}}: A  Method for Explaining Uncertainty Estimates},
+title={Getting a {CLUE}: A  Method for Explaining Uncertainty Estimates},
 author={Javier Antoran and Umang Bhatt and Tameem Adel and Adrian Weller and Jos{\'e} Miguel Hern{\'a}ndez-Lobato},
 booktitle={International Conference on Learning Representations},
 year={2021},
