### 2024-11-12 #6567 @BirkhoffG
Hi @crvernon - I have fixed these two issues. Thank you for your help in this submission.
